move codegen into jsinterpreter directory
