+++ apps/src/eval/eval.js
@@ -24,7 +24,7 @@ var ReactDOM = require('react-dom');
 var studioApp = require('../StudioApp').singleton;
 var commonMsg = require('@cdo/locale');
 var evalMsg = require('./locale');
-var codegen = require('../codegen');
+var codegen = require('../lib/tools/jsinterpreter/codegen');
 var api = require('./api');
 var Provider = require('react-redux').Provider;
 var AppView = require('../templates/AppView');
